Renamed `DH\DoctrineAuditBundle\EventSubscriber` namespace to `DH\DoctrineAuditBundle\Event`
